refactor(aigateway): split the OpenRouter discovery parsers into siblings
`plugins/openrouter_provider/discovery.py` was 497 lines against the 450-line
limit the plan binds to touched Python files. The plan names the provider
parsers as the seam, and the file already marked it with a section comment:

- `openapi_schema.py` — the OME-647 OpenAPI shape reading. Mapping the subset
  of JSON Schema the public document actually uses onto the gateway's
  deliberately small `ParameterSchema` is the largest and most self-contained
  concern in the module, and it is the half with no bearing on the catalog.
- `observations.py` — the vocabulary both parsers and the reviewed
  labelled-local inventory share: the wrapper request path, the observation
  constructor, and the dedup. These could not simply stay behind, because the
  OpenAPI half needs them and importing back from `discovery` would close a
  cycle. Declaring the three source labels here as well makes the §5.1
  distinctness requirement visible in one place instead of two.

`discovery.py` keeps the fixed public source URLs, the catalog reading, the
reviewed labelled-local evidence, the source revision, the measured OpenAPI
bounds, and the async fetch that orchestrates both documents — and re-exports
`parse_openapi_endpoint_observations`, so it stays the single import path for
both parsers.

Behaviour-preserving, and proved rather than asserted:

- Zero test files touched — the append-only gate passing unskipped is the
  machine proof.
- pyright green: every import site still resolves.
- Module surface diffed by AST against the previous revision. Nothing became
  newly reachable; all twelve names imported from `discovery` anywhere in
  `src/` or `tests/` are still present, and every source label, URL, schema
  name and the snapshot revision string are byte-identical. What left is five
  incidental import leaks and thirteen private symbols with no reference
  outside the module — `ENDPOINT_SOURCE` among them, which tests assert only
  as the literal `openrouter:openapi`.
- Enabled-OpenRouter conformance green (11 passed).

276 / 214 / 57 lines.

Refs: OME-653
